+++ src/xdot_parse/draw.rs
@@ -64,22 +64,7 @@ impl Pen {
             font_characteristics,
         }
     }
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
 
 #[cfg(feature = "pyo3")]
