+++ src/xdot_parse/draw/attrs.rs
@@ -28,22 +28,7 @@ impl Default for Rgba {
 #[cfg(feature = "pyo3")]
 #[pyo3::pymethods]
 impl Rgba {
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
 
 /// Line style for node borders and edges.
@@ -79,22 +64,7 @@ impl FromStr for Style {
 #[cfg(feature = "pyo3")]
 #[pyo3::pymethods]
 impl Style {
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
 
 bitflags! {
@@ -118,20 +88,5 @@ bitflags! {
 #[cfg(feature = "pyo3")]
 #[pyo3::pymethods]
 impl FontCharacteristics {
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
