+++ src/xdot_parse/shapes.rs
@@ -29,22 +29,7 @@ impl Ellipse {
     fn new(x: f32, y: f32, w: f32, h: f32, filled: bool) -> Self {
         Ellipse { x, y, w, h, filled }
     }
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
 impl From<Ellipse> for Shape {
     fn from(val: Ellipse) -> Self {
@@ -86,22 +71,7 @@ impl Points {
             filled,
         }
     }
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
 impl From<Points> for Shape {
     fn from(val: Points) -> Self {
@@ -146,22 +116,7 @@ impl Text {
             text,
         }
     }
-    // TODO: dedup
-    fn __richcmp__(
-        &self,
-        other: &Self,
-        op: pyo3::pyclass::CompareOp,
-        py: pyo3::Python,
-    ) -> pyo3::PyObject {
-        use pyo3::pyclass::CompareOp::*;
-        use pyo3::IntoPy;
-
-        match op {
-            Eq => (self == other).into_py(py),
-            Ne => (self != other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
+    impl_richcmp_eq!();
 }
 impl From<Text> for Shape {
     fn from(val: Text) -> Self {
